+++ lib/pundit/cache_store/legacy_store.rb
@@ -2,12 +2,19 @@
 
 module Pundit
   module CacheStore
+    # A cache store that uses only the record as a cache key, and ignores the user.
+    #
+    # The original cache mechanism used by Pundit.
+    #
     # @api private
     class LegacyStore
       def initialize(hash = {})
         @store = hash
       end
 
+      # A cache store that uses only the record as a cache key, and ignores the user.
+      #
+      # @note `nil` results are not cached.
       def fetch(user:, record:)
         _ = user
         @store[record] ||= yield