+++ lib/pundit/rspec.rb
@@ -1,13 +1,29 @@
 # frozen_string_literal: true
 
+# Array#to_sentence
+require "active_support/core_ext/array/conversions"
+
 module Pundit
+  # Namespace for Pundit's RSpec integration.
   module RSpec
+    # Namespace for Pundit's RSpec matchers.
     module Matchers
       extend ::RSpec::Matchers::DSL
 
+      # @!method description=(description)
       class << self
+        # Used to build a suitable description for the Pundit `permit` matcher.
+        # @api public
+        # @param value [String, Proc]
+        # @example
+        #   Pundit::RSpec::Matchers.description = ->(user, record) do
+        #     "permit user with role #{user.role} to access record with ID #{record.id}"
+        #   end
         attr_writer :description
 
+        # Used to retrieve a suitable description for the Pundit `permit` matcher.
+        # @api private
+        # @private
         def description(user, record)
           return @description.call(user, record) if defined?(@description) && @description.respond_to?(:call)
 
@@ -32,15 +48,21 @@ module Pundit
         end
 
         failure_message_proc = lambda do |policy|
-          was_were = @violating_permissions.count > 1 ? "were" : "was"
           "Expected #{policy} to grant #{permissions.to_sentence} on " \
-          "#{record} but #{@violating_permissions.to_sentence} #{was_were} not granted"
+          "#{record} but #{@violating_permissions.to_sentence} #{was_or_were} not granted"
         end
 
         failure_message_when_negated_proc = lambda do |policy|
-          was_were = @violating_permissions.count > 1 ? "were" : "was"
           "Expected #{policy} not to grant #{permissions.to_sentence} on " \
-          "#{record} but #{@violating_permissions.to_sentence} #{was_were} granted"
+          "#{record} but #{@violating_permissions.to_sentence} #{was_or_were} granted"
+        end
+
+        def was_or_were
+          if @violating_permissions.count > 1
+            "were"
+          else
+            "was"
+          end
         end
 
         description do
@@ -53,21 +75,57 @@ module Pundit
           failure_message(&failure_message_proc)
           failure_message_when_negated(&failure_message_when_negated_proc)
         else
+          # :nocov:
+          # Compatibility with RSpec < 3.0, released 2014-06-01.
           match_for_should(&match_proc)
           match_for_should_not(&match_when_negated_proc)
           failure_message_for_should(&failure_message_proc)
           failure_message_for_should_not(&failure_message_when_negated_proc)
+          # :nocov:
+        end
+
+        if ::RSpec.respond_to?(:current_example)
+          def current_example
+            ::RSpec.current_example
+          end
+        else
+          # :nocov:
+          # Compatibility with RSpec < 3.0, released 2014-06-01.
+          def current_example
+            example
+          end
+          # :nocov:
         end
 
         def permissions
-          current_example = ::RSpec.respond_to?(:current_example) ? ::RSpec.current_example : example
-          current_example.metadata[:permissions]
+          current_example.metadata.fetch(:permissions) do
+            raise KeyError, <<~ERROR.strip
+              No permissions in example metadata, did you forget to wrap with `permissions :show?, ...`?
+            ERROR
+          end
         end
       end
       # rubocop:enable Metrics/BlockLength
     end
 
+    # Mixed in to all policy example groups to provide a DSL.
     module DSL
+      # @example
+      #   describe PostPolicy do
+      #     permissions :show?, :update? do
+      #       it { is_expected.to permit(user, own_post) }
+      #     end
+      #   end
+      #
+      # @example focused example group
+      #   describe PostPolicy do
+      #     permissions :show?, :update?, :focus do
+      #       it { is_expected.to permit(user, own_post) }
+      #     end
+      #   end
+      #
+      # @param list [Symbol, Array<Symbol>] a permission to describe
+      # @return [void]
       def permissions(*list, &block)
         metadata = { permissions: list, caller: caller }
 
@@ -81,6 +139,9 @@ module Pundit
       end
     end
 
+    # Mixed in to all policy example groups.
+    #
+    # @private not useful
     module PolicyExampleGroup
       include Pundit::RSpec::Matchers