+++ spec/policy_finder_spec.rb
@@ -2,12 +2,16 @@
 
 require "spec_helper"
 
-class Foo; end
 RSpec.describe Pundit::PolicyFinder do
   let(:user) { double }
   let(:post) { Post.new(user) }
   let(:comment) { CommentFourFiveSix.new }
   let(:article) { Article.new }
+
+  describe "SUFFIX" do
+    specify { expect(described_class::SUFFIX).to eq "Policy" }
+    specify { expect(Pundit::SUFFIX).to eq(described_class::SUFFIX) }
+  end
 
   describe "#scope" do
     subject { described_class.new(post) }