+++ spec/support/models/customer/post.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Customer
+  class Post < ::Post
+    extend ActiveModel::Naming
+
+    def self.policy_class
+      PostPolicy
+    end
+  end
+end