+++ spec/support/models/dummy_current_user.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class DummyCurrentUser
+  def update?
+    user
+  end
+end